Move error logger to new file
