+++ server/src/constants.ts
@@ -35,3 +35,4 @@ export const NOTIFICATION_CHANGE_MARK = 'changeMark';
 export const NOTIFICATION_CHANGE_OFFER_STATUS = 'changeOfferStatus';
 export const NEW_MESSAGE = 'newMessage';
 export const CHANGE_BLOCK_STATUS = 'CHANGE_BLOCK_STATUS';
+export const LOG_PATH = path.resolve(__dirname, 'logs/latest.log');
